+++ app/static/lib/control-menu.js
@@ -3,6 +3,9 @@ import * as icon from './../css/icons.js';
 import {
   fontList
 } from './main.js';
+import {
+  svgNameSpace
+} from './dom-utils.js'
 
 export function createControlsMenu(parentElement, scale) {
 
@@ -310,6 +313,12 @@ time with large files`;
   let imagePanel = document.createElement('div');
   imagePanel.id = 'image-panel';
   imagePanel.style.display = 'none';
+  var svg = document.createElementNS(svgNameSpace, 'svg');
+  svg.id = 'source-image-container';
+  var g = document.createElementNS(svgNameSpace, 'svg');
+  g.id = 'source-image-svg';
+  svg.appendChild(g);
+  imagePanel.append(svg);
   pixPanel.appendChild(imagePanel);
 
 } // createControlsMenu()
@@ -329,25 +338,30 @@ export function setBreaksOptions(tkAvailableOptions, defaultValue = 'auto') {
   if (defaultValue == '') defaultValue = 'auto';
   let breaksEl = document.getElementById('breaks-select');
   var breaksOpts = {
+    none: 'None',
     auto: 'Automatic',
-    smart: 'Smart',
+    // measure: 'Measure',
     line: 'System',
     encoded: 'System and page',
-    none: 'None'
+    smart: 'Smart'
   };
-  var breaks = utils.findKey('breaks', tkAvailableOptions);
-  for (let index of breaks.values) {
-    if (breaksOpts[index]) {
-      breaksEl[breaksEl.options.length] = new Option(breaksOpts[index], index);
-      if (index == defaultValue) {
-        breaksEl[breaksEl.options.length - 1].selected = 'selected';
-      }
-    } else {
-      breaksEl[breaksEl.options.length] = new Option(index, index);
-    }
-    // disable breaks=smart by default; only enabled with speedMode=false
-    if (index === 'smart') breaksEl[breaksEl.length - 1].disabled = true;
+  for (let key of Object.keys(breaksOpts)) {
+    breaksEl[breaksEl.options.length] = new Option(breaksOpts[key], key);
+    if (key === 'smart') breaksEl[breaksEl.length - 1].disabled = true;
   }
+  // var breaks = utils.findKey('breaks', tkAvailableOptions);
+  // for (let index of breaks.values) {
+  //   if (breaksOpts[index]) {
+  //     breaksEl[breaksEl.options.length] = new Option(breaksOpts[index], index);
+  //     if (index == defaultValue) {
+  //       breaksEl[breaksEl.options.length - 1].selected = 'selected';
+  //     }
+  //   } else {
+  //     breaksEl[breaksEl.options.length] = new Option(index, index);
+  //   }
+  // disable breaks=smart by default; only enabled with speedMode=false
+  //   if (index === 'smart') breaksEl[breaksEl.length - 1].disabled = true;
+  // }
 }
 
 export function handleSmartBreaksOption(speedMode) {